requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-cleanup==3.2.0
django-storages==1.7.1
django-widget-tweaks==1.4.3
docopt==0.6.2
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gunicorn==19.9.0
idna==2.8
mysql-connector-python==8.0.16
mysqlclient==1.4.2
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wincertstore==0.2
yarg==0.1.9
